QFuture: fix result reporting in the filter mode
After inserting a result into the result store, a wrong index for end
is used for reporting the range of ready results. This has been fixed
in Qt 6 in scope of 44ceb56455c82df3e6b1c9a2fa373cac14a039f8 already.
The problem was noticed also for Qt 5 long time ago (see the comments
of QTBUG-6799), so port the fix to 5.15. The relevant part for testing
the fix is taken from the same patch.

Fixes: QTBUG-83182
Change-Id: I99121732576befa9ec32f758746d9a22a30e4261
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>