testlib: Run test result expectation tests with stdout as output
For historical reasons a few of the subtests are skipped when
running with anything but the plain text logger to stdout.

To ensure we have as broad test coverage as possible for the
expected output of the various loggers we run these tests in
stdout-mode.

Change-Id: I856905d1543afe89710533657a55bd599c0305fd
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit cba2adb1952db08202e3c7a633cfa3f62575fb99)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>